==> 1 - assessing.ipynb <==
## Avaliando
Use o espaço abaixo para explorar os arquivos all_alpha_08.csv e all_alpha_18.csv para responder as perguntas do teste abaixo.

import pandas as pd

# Importando os dados de 2008
df_08 = pd.read_csv('all_alpha_08.csv')
df_18 = pd.read_csv('all_alpha_18.csv')

print("df_08")
display(df_08.head())
print("df_18")
display(df_18.head())

# número de amostras em cada conjunto
print("Numero de amostras - 08: {0} - 18: {1}".format(
        df_08.shape[0], df_18.shape[0]))

# número de colunas em cada conjunto
print("Numero de colunas - 08: {0} - 18: {1}".format(
        df_08.shape[1], df_18.shape[1]))

# linhas duplicadas em cada conjunto
print("Duplicadas - 08: {0} - 18: {1}".format(
        df_08.duplicated().sum(), df_18.duplicated().sum()))

df_08.info()

df_18.info()

def check_types(df):
    for column in df.columns:
        print("coluna: {0}, tipo: {1}".format(column, type(df[column][0])))
# tipos de dados das colunas
check_types(df_08)

# tipos de dados das colunas
check_types(df_18)

# número de valores únicos não nulos para atributos em cada conjunto
df_08.nunique()

# número de valores únicos não nulos para atributos em cada conjunto
df_18.nunique()

# o que esses valores únicos são e quanto valem cada
df_08["Fuel"].unique()

# o que esses valores únicos são e quanto valem cada
df_18["Fuel"].unique()
