==> 4 - inspect_datatypes.ipynb <==
# Inspecionando tipos de dados
Use o espaço abaixo para explorar os arquivos `data_08.csv` e `data_18.csv` para responder às perguntas do teste abaixo sobre tipos de dados.

import pandas as pd

df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

#Quais das mudanças a seguir podem ser feitas à coluna “cyl.” nos dois conjuntos de dados de forma clara e consistente?
print("Cyl em 08:{0} \nCyl em 18: {1}".format(
        type(df_08['cyl'][0]), type(df_18['cyl'][0])))

# Quais das mudanças a seguir podem ser feitas à coluna “air_pollution_score” nos dois conjuntos de dados de forma clara e consistente?
print("Air Pollution em 08:{0} \nAir Pollution em 18: {1}".format(
        type(df_08['air_pollution_score'][0]), type(df_18['air_pollution_score'][0])))



# Quais dos atributos a seguir precisam ser transformados de strings em floats?
print("City Mpg em 08:{0} \nCity Mpg em 18: {1}".format(
        type(df_08['city_mpg'][0]), type(df_18['city_mpg'][0])))

print("Hwy Mpg em 08:{0} \nHwy Mpg em 18: {1}".format(
        type(df_08['hwy_mpg'][0]), type(df_18['hwy_mpg'][0])))

print("Cmb Mpg em 08:{0} \nCmb Mpg em 18: {1}".format(
        type(df_08['cmb_mpg'][0]), type(df_18['cmb_mpg'][0])))

# Quais mudanças precisam ser feitas para deixar a coluna “greenhouse_gas_score” consistente nos dois conjuntos de dados?
print("Greenhouse Gas em 08:{0} \nGreenhouse Gas em 18: {1}".format(
        type(df_08['greenhouse_gas_score'][0]), type(df_18['greenhouse_gas_score'][0])))
