==> 6 - fix_datatypes_air_pollution.ipynb <==
# Corrigindo o tipo de dados de `air_pollution_score`
- 2008: converta de string para float
- 2018: converta de int para float

# carregue os conjuntos de dados
import pandas as pd

df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

# tente usar as funções to_numeric ou astype do Pandas para converter a coluna
# air_pollution_score do conjunto de dados de 2008 para float -- isso não vai funcionar

#df_08['air_pollution_score'] = df_08['air_pollution_score'].astype(int)
df_08['air_pollution_score'] = pd.to_numeric(df_08['air_pollution_score'])

# Resolvendo a questão
Parece que isso não vai ser tão simples quanto converter o tipo de dados. Segundo o erro abaixo, o valor na linha 582 seria "6/4" - vamos dar uma olhada.

df_08.iloc[582]

# Não é só o nível de poluição do ar!
As colunas mpg e os nível de gases de efeito estufa também parecem ter o mesmo problema - talvez seja por isso que todos eles foram salvos como strings! Segundo [este link](http://www.fueleconomy.gov/feg/findacarhelp.shtml#airPollutionScore), que eu encontrei na documentação em PDF:

    "Se um veículo pode operar com mais de um tipo de combustível, uma estimativa é fornecida para cada tipo de combustível."
    
Ahh.. então todos veículos com mais de um tipo de combustível, ou híbridos, como aquele visto acima (que usa etanol E gás) terão uma string que armazena dois valores - um para cada. Isto é um pouco complicado, então eu vou mostrar como fazê-lo com o conjunto de dados de 2008, para você então tentar fazer o mesmo com o conjunto de dados de 2018.

# Primeiro, vamos obter todos os híbridos de 2008
hb_08 = df_08[df_08['fuel'].str.contains('/')]
hb_08

Parece que este conjunto de dados só tem um! O de 2018 tem MUITOS mais - mas não se preocupe - os passos que eu estou fazendo aqui vão funcionar para este também!

# híbridos em 2018
hb_18 = df_18[df_18['fuel'].str.contains('/')]
hb_18

Vamos pegar cada linha que contém um carro híbrido e dividí-la em duas novas linhas - uma com os valores do primeiro tipo de combustível (valores antes da "/") e outra com os valores do segundo tipo de combustível (valores depois da "/"). Vamos separá-los com dois dataframes!

# crie duas cópias do dataframe com híbridos de 2008
df1 = hb_08.copy()  # data on first fuel type of each hybrid vehicle
df2 = hb_08.copy()  # data on second fuel type of each hybrid vehicle

# Cada um deve se parecer com isto
df1

Para a próxima parte, vamos usar a função apply do Pandas. Veja a documentação [aqui](https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.apply.html).

# colunas que devem ser divididas por "/"
split_columns = ['fuel', 'air_pollution_score', 'city_mpg', 'hwy_mpg', 'cmb_mpg', 'greenhouse_gas_score']

# aplique a função split para cada coluna de cada cópia do dataframe
for c in split_columns:
    df1[c] = df1[c].apply(lambda x: x.split("/")[0])
    df2[c] = df2[c].apply(lambda x: x.split("/")[1])

# este dataframe armazena informações sobre o PRIMEIRO tipo de combustível do híbrido
# i.e. os valores antes das "/"s
df1

# este dataframe armazena informações sobre o SEGUNDO tipo de combustível do híbrido
# i.e. os valores depois das "/"s
df2

# combine dataframes para adicionar ao dataframe original
new_rows = df1.append(df2)

# agora temos linhas separadas para cada tipo de combustível de cada veículo!
new_rows

# descarte as linhas originais que continham híbridos
df_08.drop(hb_08.index, inplace=True)

# adicione as novas linhas separadas
df_08 = df_08.append(new_rows, ignore_index=True)

# verifique se todas linhas originais com híbridos, que continham "/"s, sumiram
df_08[df_08['fuel'].str.contains('/')]

df_08.shape

# Repita este processo para o conjunto de dados de 2018

# crie duas cópias do dataframe com híbridos de 2018, hb_18
df1 = hb_18.copy()  # data on first fuel type of each hybrid vehicle
df2 = hb_18.copy()  # data on second fuel type of each hybrid vehicle


### Divida valores para `fuel`, `city_mpg`, `hwy_mpg`, `cmb_mpg`
Você não precisa dividir as colunas `air_pollution_score` ou `greenhouse_gas_score` aqui porque estas colunas já são ints no conjunto de dados de 2018.

# lista de colunas que devem ser divididas
split_columns = ['fuel', 'city_mpg', 'hwy_mpg', 'cmb_mpg']

# aplique a função split para cada coluna de cada cópia do dataframe
for c in split_columns:
    df1[c] = df1[c].apply(lambda x: x.split("/")[0])
    df2[c] = df2[c].apply(lambda x: x.split("/")[1])

# combine os dois dataframes
new_rows = df1.append(df2)

# descarte cada linha original que continha híbridos no dataframe original de 2018
# faça isso usando a função drop do Pandas com o índice do hb_18
df_18.drop(hb_18.index, inplace=True)

# anexe new_rows a df_18
df_18 = df_18.append(new_rows, ignore_index=True)

# verifique se elas sumiram
df_18[df_18['fuel'].str.contains('/')]

df_18.shape

### Agora podemos confortavelmente continuar as mudanças necessárias para `air_pollution_score`! Aqui vão elas novamente:
- 2008: converta de string para float
- 2018: converta de int para float

# converta a coluna de poluição do ar de 2008 de string para float
df_08['air_pollution_score'] = df_08['air_pollution_score'].astype(float)

# converta a coluna de poluição do ar de 2018 de int to float
df_18['air_pollution_score'] = df_08['air_pollution_score'].astype(float)

df_08.to_csv('data_08.csv', index=False)
df_18.to_csv('data_18.csv', index=False)
