==> 7 - fix_datatypes_mpg_greenhouse.ipynb <==
## Corrija os tipos de dados de `city_mpg`, `hwy_mpg`, `cmb_mpg`
    2008 e 2018: converta de string para float

# carregue os conjuntos de dados
import pandas as pd

df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

# converta colunas mpg para floats
mpg_columns = ['city_mpg', 'hwy_mpg', 'cmb_mpg']
for c in mpg_columns:
    df_18[c] = df_18[c].astype(float)
    df_08[c] = df_08[c].astype(float)

## Corrija o tipo de dados `greenhouse_gas_score`
    2008: converta de float para int

# converta de float para int
df_08['greenhouse_gas_score'] = df_08['greenhouse_gas_score'].astype(int)

## Todos tipos de dados agora estão corretos! Verifique uma última vez para confirmar todas mudanças.

df_08.dtypes

df_18.dtypes

df_08.dtypes == df_18.dtypes

# Salve seus novos conjuntos de dados LIMPOS como novos arquivos!
df_08.to_csv('clean_08.csv', index=False)
df_18.to_csv('clean_18.csv', index=False)
