==> 8 - drawing_conclusions.ipynb <==
# Tirando conclusões
Use o espaço abaixo para resolver questões sobre os conjuntos de dados `clean_08.csv` e `clean_18.csv`

# carregue os conjuntos de dados
import pandas as pd

df_08 = pd.read_csv('clean_08.csv')
df_18 = pd.read_csv('clean_18.csv')

### Q1: Existem mais modelos únicos usando fontes alternativas de combustível? Em qual proporção?

df_08['fuel'].unique()

fuel_08 = df_08['fuel'].unique()
fuel_18 = df_18['fuel'].unique()

def calc_fuel_prop(df):
    fuels = df['fuel'].unique()
    for fuel in fuels:
        total = df.shape[0] # qtd total de registros
        proporção = df.query("fuel == @fuel")['fuel'].count()/total
        print(f'Combustível: {fuel} Quantidade {proporção}')
print('08')
calc_fuel_prop(df_08)
print('18')
calc_fuel_prop(df_18)

### Q2: Qual foi o nível de melhoria das classes de veículo no quesito "economia de combustível"?

def calc_veh_fuel(df1, df2):
    classes = df2['veh_class'].unique()
    for v_class in classes:
        print('Classe: {0} Variação: {1}'.format(
                v_class, df1.query("veh_class == @v_class")['veh_class'].count()/
                         df2.query("veh_class == @v_class")['veh_class'].count()))

calc_veh_fuel(df_08, df_18)

### Q3: Quais são as características dos veículos SmartWay? Elas mudaram com o tempo?

df_08.groupby("smartway").describe()

df_18.groupby("smartway").describe()

### Q4: Quais características estão associadas à uma maior economia de combustível?

df_08.corr()

df_18.corr()
