==> 10 - merging_datav.ipynb <==
# Mesclando conjuntos de dados
Use merges do Pandas para criar um conjunto de dados combinado a partir de `clean_08.csv` e `clean_18.csv`.

# carregue conjuntos de dados
import pandas as pd

df_08 = pd.read_csv('clean_08.csv')
df_18 = pd.read_csv('clean_18.csv')

### Crie conjunto de dados combinado

# renomeie colunas de 2008
df_08.rename(columns=lambda x: x[:10] + "_2008", inplace=True)

# exiba para verificar nomes
df_08.head()

# mescle os conjuntos de dados
df_combined = pd.concat([df_08, df_18], axis=1, join='inner')


# exiba para verificar a mesclagem
df_combined.head()

Salve o conjunto de dados combinado

df_combined.to_csv('combined_dataset.csv', index=False)
